read_text

Read .txt, .tab, .data (text) file.

Usage

read_text(
  filename,
  delimiter = NULL,
  first_column_names = NULL,
  dtype = "float32"
)

Arguments

filename

Data file, filename or stream.

delimiter

Delimiter that separates data within text file. If NULL, will split at arbitrary number of white spaces, which is different from enforcing splitting at single white space ' '.

first_column_names

Assume the first column stores row names.

dtype

Numpy data type.

Details

Same as read_csv() but with default delimiter NULL.

Examples

if (FALSE) { # \dontrun{
ad <- read_text("matrix.tab")
} # }